## Description
```
Added posture bar ui. 
Cleaned up and ensured proper posture bar rendering.
```

## Changes Made
```
Outline the specific changes made in this pull request. (if there are too many to list you should've made multiple PRs)
```

## Screenshots (if applicable)
```
![image](https://github.students.cs.ubc.ca/CPSC427-2024W-T1/Team25/assets/22088/1f29073f-1503-4434-9bed-9b1b7ec98b7d)

```

## Testing
```
Describe the testing that has been done to ensure the changes are working as expected.
```

### Bugs
```
List known bugs for your changes
```

## Checklist 
```
Both the requestee and the reviewer should review this checklist
```
- [ ] Code Compiles on your machine
- [ ] PR Filled in (remove the code formatted blocks)
- [ ] Code is appropriately compartmentalized with 
- [ ] feature has no/minimal interference for other functions
- [ ] Code is well documented with comments
- [ ] Code is readable
- [ ] Tests (if included) pass

## Additional Notes
```
Add any additional notes or context that may be helpful for reviewers.
```

+++ src/renderer.hpp
@@ -142,8 +142,8 @@ public:
     void handleP1Health(float p1Health);
     void handleP2Health(float p2Health);
 
-    void handleP1Posture(float p1Posture);
-    void handleP2Posture(float p2Posture);
+    void handleP1Posture(int p1Posture);
+    void handleP2Posture(int p2Posture);
 
     void renderSelectorTriangleP1(float x, float y, float width, float height, bool p1);
     void renderSelectorTriangleP2(float x, float y, float width, float height, bool p2);
